move db exceptions to db package
