+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/db/MissingRelationBetweenTables.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.orderlyweb.models
+package org.vaccineimpact.orderlyweb.db
 
 class MissingRelationBetweenTables(from: org.jooq.Table<*>, to: org.jooq.Table<*>) : Exception(
         "Attempted to construct join from ${from.name} to ${to.name}, but there are no keys relating those tables."
